--- starknet_py/tests/e2e/client/conftest.py
@@ -1,3 +1,0 @@
-import os
-
-directory = os.path.dirname(__file__)
