+++ starknet_py/tests/e2e/client/fixtures/get_class_hash.sh
